Progression
A lot of progression is complete.
There is a lot that I did not test well here yet, especially revolving around the many to many relationships in ef core. I need to go through the db context and register my relationships better.
I know I am going to have issues in certain places (especially updates and deletes) but I am planning on knocking those out as I develop the UI.
I did not just commit that thinking that it is 100% working, I know I am going to have to clean up some of that stuff. I'm okay with that.